Fix ts errors
