+++ packages/react-learn-card/src/components/CertificateDisplayCard/index.ts
@@ -2,4 +2,15 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from "./CertificateBackFace";
+export * from "./CertificateCornerIcon";
 export * from "./CertificateDisplayCard";
+export * from "./CertificateFrontFace";
+export * from "./CertificateImageDisplay";
+export * from "./CertificateProfileImageDisplay";
+export * from "./CertificateRibbon";
+export * from "./InfoBox";
+export * from "./MediaAttachmentsBox";
+export * from "./TruncateTextBox";
+export * from "./VerificationRow";
+export * from "./VerificationsBox";
