+++ packages/react-learn-card/src/components/VCDisplayCard2/VCDisplayCard2.tsx
@@ -16,8 +16,7 @@ import {
     getColorForVerificationStatus,
     getInfoFromCredential,
 } from '../../helpers/credential.helpers';
-import { BoostAchievementCredential, IssueHistory, LCCategoryEnum } from '../../types';
-import { MediaMetadata, VideoMetadata } from './MediaAttachmentsBox';
+import { BoostAchievementCredential, IssueHistory, LCCategoryEnum, MediaMetadata, VideoMetadata  } from '../../types';
 import { CertificateDisplayCard } from '../CertificateDisplayCard';
 
 export type CredentialIconType = {
