+++ packages/react-learn-card/src/components/VCDisplayCard2/index.ts
@@ -7,11 +7,15 @@ export * from "./InfoBox";
 export * from "./IssueHistoryBox";
 export * from "./MediaAttachmentsBox";
 export * from "./RibbonEnd";
+export * from "./SelectedSkills";
 export * from "./SkillsBox";
+export * from "./Subskill";
+export * from "./TopLevelSkill";
 export * from "./TruncateTextBox";
 export * from "./VC2BackFace";
 export * from "./VC2FrontFaceInfo";
 export * from "./VCDisplayCard2";
 export * from "./VCDisplayCardCategoryType";
+export * from "./VCDisplayCardSkillsCount";
 export * from "./VerificationRow";
 export * from "./VerificationsBox";
