+++ packages/react-learn-card/src/components/svgs/index.ts
@@ -2,6 +2,8 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from "./AccommodationsIcon";
+export * from "./AccomplishmentsIcon";
 export * from "./AchievementWithTitle";
 export * from "./AcuteCheckmark";
 export * from "./AwardRibbon";
@@ -12,19 +14,30 @@ export * from "./CircleWithText";
 export * from "./Coins";
 export * from "./DownRightArrow";
 export * from "./ExclamationPoint";
+export * from "./ExperienceIcon";
+export * from "./FatArrow";
 export * from "./GenericArrow";
 export * from "./GenericDocumentIcon";
 export * from "./Graduation";
 export * from "./InfoIcon";
+export * from "./IssuerSeal";
 export * from "./KeyIcon";
 export * from "./LeftArrow";
 export * from "./Lightbulb";
+export * from "./Line";
 export * from "./LinkIcon";
 export * from "./PdfIcon";
+export * from "./PersonBadge";
 export * from "./PuzzlePiece";
+export * from "./RedFlag";
+export * from "./RoundedCorner";
 export * from "./RoundedX";
 export * from "./ShieldChevron";
+export * from "./Smiley";
+export * from "./SocialBadgesIcon";
 export * from "./Trophy";
+export * from "./UnknownVerifierBadge";
 export * from "./User";
+export * from "./VerifiedBadge";
 export * from "./VideoIcon";
 export * from "./X";
